# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectConfigTokenResponse": {
        "additionalProperties": true,
        "properties": {
          "config_token": {
            "title": "Config Token",
            "type": "string"
          },
          "project": {
            "title": "Project",
            "type": "string"
          }
        },
        "required": [
          "project",
          "config_token"
        ],
        "title": "ProjectConfigTokenResponse",
        "type": "object"
      },
      "ProjectDisplayNameUpdate": {
        "properties": {
          "display_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "ProjectDisplayNameUpdate",
        "type": "object"
      },
...
      "QueuedJobResponse": {
        "additionalProperties": true,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "is_idempotent": {
            "title": "Is Idempotent",
            "type": "boolean"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "progress": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "project": {
            "title": "Project",
            "type": "string"
          },
          "project_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Uuid"
          },
          "queue_position": {
            "title": "Queue Position",
            "type": "integer"
          },
          "retry_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Of"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Uuid"
          },
          "total_steps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Steps"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "job_id",
          "project",
          "project_uuid",
          "tenant_uuid",
          "created_by",
          "action",
          "status",
          "message",
          "progress",
          "current_step",
          "total_steps",
          "started_at",
          "finished_at",
          "error_code",
          "is_idempotent",
          "retryable",
          "retry_of",
          "attempt",
          "created_at",
          "updated_at",
          "queue_position"
        ],
        "title": "QueuedJobResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Control plane for isolated Supabase projects on shared infrastructure.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Supabase Multitenant Projects API",
    "version": "0.13.0-alpha"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{project_name}": {
      "delete": {
        "operationId": "delete_project_api_projects__project_name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Step-Up-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Step-Up-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/projects/{project_name}/config-token": {
      "get": {
        "description": "Entrega o token compartilhado aos membros do projeto e registra a leitura.",
        "operationId": "get_project_config_token_api_projects__project_name__config_token_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectConfigTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Config Token",
        "tags": [
          "project-rename"
        ]
      }
    },
    "/api/projects/{project_name}/display-name": {
      "patch": {
        "description": "Atualiza apenas o display_name do projeto (sem migrar infraestrutura).",
        "operationId": "update_project_display_name_api_projects__project_name__display_name_patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDisplayNameUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateDisplayNameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project Display Name",
        "tags": [
          "project-rename"
        ]
      }
    },
...
    "/api/projects/{project_name}/queue-status": {
      "get": {
        "description": "Retorna o estado atual da fila de a\u00e7\u00f5es do projeto.\n\nInclui o job em execu\u00e7\u00e3o (se houver), o tamanho da fila, e os jobs\npendentes/rodando do banco para fins de UI (polling).",
        "operationId": "get_project_queue_status_api_projects__project_name__queue_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectQueueStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Queue Status",
        "tags": [
          "project-rename"
        ]
      }
    },
...
    "/api/projects/{project_name}/rename": {
      "post": {
        "description": "Renomeia o slug/path do projeto (migra\u00e7\u00e3o completa em background).\n\nO escopo inclui: nome interno na meta DB, banco Postgres, roles\npor projeto, replication slots do Realtime, tenant Supavisor,\ndiret\u00f3rio f\u00edsico e templates (nginx, docker-compose, .env).",
        "operationId": "rename_project_api_projects__project_name__rename_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenameProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Project",
        "tags": [
          "project-rename"
        ]
      }
    },
    "/api/projects/{project_name}/rename-history": {
      "get": {
        "description": "Retorna auditoria e historico duravel de nome/path do projeto.",
        "operationId": "get_project_rename_history_api_projects__project_name__rename_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRenameHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Rename History",
        "tags": [
          "project-rename"
        ]
      }
    },
...
  }
}